requirements.txt
aiogram==3.12.0
python-dotenv==1.0.1
google-genai==1.32.0
faiss-cpu==1.12.0
PyPDF2==3.0.1
numpy==2.1.3
